Move ovnk code to package cni
